+++ RNSentryCocoaTester/RNSentryCocoaTesterTests/RNSentryCocoaTesterTests-Bridging-Header.h
@@ -1,0 +1,1 @@
+#import "RNSentryBreadcrumb.h"